refactor(data-model): the membership predicates get a module of their own
`validity-check.ts` holds the predicates deciding whether a value belongs to
the `FabricValue` type -- `isValidFabricValueLayer()`, its throwing form
`assertValidFabricValueLayer()`, `isValidFabricValue()`,
`isValidFabricPlainObject()`, and `isValidFabricNativeObject()` -- along with
the two helpers the throwing form uses. `type-check.ts` keeps the narrowings
asked of a value already typed as a `FabricValue`, and the `isKeyable*` and
`isWalkable*` pairs. Each unit moves byte-for-byte, and the test file splits
the same way, block for block.

Callers, the barrel, and the prose citing the old location are retargeted.
