+++ docs/specs/space-model-formal-spec/1-fabric-values.md
@@ -3952,7 +3952,7 @@ returned value is always a valid `FabricValue` regardless of its frozen state.
  * and allocation).
  *
  * Relationship to other functions and checks:
- * - `isValidFabricValue(x)` (in `type-check.ts`): the narrower check — "is `x`
+ * - `isValidFabricValue(x)` (in `validity-check.ts`): the narrower check — "is `x`
  *   already a `FabricValue`?" — which does NOT accept raw native types like
  *   `Error` or `Map`.
  * - `isValidFabricConvertibleValue(x)`: "Could `x` be converted to a
