+++ packages/cli/lib/shuttle/verbs.ts
@@ -2287,7 +2287,7 @@ const THE_VALUE_ITSELF = "the value itself";
  *
  * What makes the list closed is that it is read off the admission test rather
  * than composed beside it. `isValidFabricValueLayer`
- * (`packages/data-model/src/type-check.ts`) switches on `typeof` and, for an
+ * (`packages/data-model/src/validity-check.ts`) switches on `typeof` and, for an
  * `object`, branches four ways; this walk answers the same arms:
  *
  * - `undefined`, `bigint` and a registry-interned `symbol` are admitted and
