+++ packages/data-model/src/index.ts
@@ -35,20 +35,23 @@ export {
 } from "./value-clone.ts";
 
 export {
-  assertValidFabricValueLayer,
   isFabricContainerValue,
   isFabricObjectOrArray,
   isFabricPlainContainer,
   isFabricPlainObject,
   isKeyableObjectNotArray,
   isKeyableObjectOrArray,
+  isWalkableObjectNotArray,
+  isWalkableObjectOrArray,
+} from "./type-check.ts";
+
+export {
+  assertValidFabricValueLayer,
   isValidFabricNativeObject,
   isValidFabricPlainObject,
   isValidFabricValue,
   isValidFabricValueLayer,
-  isWalkableObjectNotArray,
-  isWalkableObjectOrArray,
-} from "./type-check.ts";
+} from "./validity-check.ts";
 
 export {
   fabricFromNativeValue,
