+++ packages/data-model/src/native-conversion.ts
@@ -5,7 +5,7 @@
  *
  * The inbound work splits along one question -- does conversion produce a new
  * value? Minting a native object's fabric form is one function, and vetting a
- * value that needs no minting is the other, in `type-check.ts`. The shallow
+ * value that needs no minting is the other, in `validity-check.ts`. The shallow
  * conversion is those two asked in that order, plus a frozenness adjustment,
  * so that a caller can ask either without having to work the answer back out
  * of what it was handed.
@@ -51,7 +51,7 @@ import { tagFromNativeValue } from "./native-type-tags.ts";
 import {
   assertValidFabricValueLayer,
   isValidFabricNativeObject,
-} from "./type-check.ts";
+} from "./validity-check.ts";
 import { cloneHelper } from "./value-clone.ts";
 import { isValidDeepFrozenFabricValue } from "./deep-freeze.ts";
